WIP: refactor: moved the files opener to services
