+++ pydatacuration/checker/services/files_opener.py
